Add files via upload
Scripts sim1.R, sim3.R and sim5.R produce the results displayed in table 1 for the parametric tests. 
Scripts sim2.R, sim4.R and sim6.R produce the results displayed in table 2 for the nonparametric tests. 
Scripts sim7.R, sim8.R and sim9.R produce the results displayed in table three for the two-stage procedure. 

To produce the plots of the paper, run script plots.R, which includes the results of all simulations one to nine, because each of these may run from several hours to multiple days depending on the machine.